requirements.txt
numpy==1.19.5
streamlit==0.68.1
pandas==1.0.1
qrcode==6.1
Pillow==8.1.0
